refactor: fix useUserSelector
