+++ project-management-app/src/store/selectors.ts
@@ -1,4 +1,4 @@
 import { useSelector } from 'react-redux';
 import { TStore } from '.';
 
-export const useUserSelector = () => useSelector((store: TStore) => store.userReducer.user);
+export const useUserSelector = () => useSelector((store: TStore) => store.userReducer);
