+++ project-management-app/src/store/slices/user.ts
@@ -1,6 +1,4 @@
 import { createSlice, PayloadAction } from '@reduxjs/toolkit';
-import { useSelector } from 'react-redux';
-import { TStore } from '..';
 import { AuthorizedUser } from '../../interfaces/user';
 import { getFromLocalStorage } from '../../utils/common';
 
@@ -21,8 +19,6 @@ const userSlice = createSlice({
     },
   },
 });
-
-export const useUserSelector = () => useSelector((state: TStore) => state.userReducer.user);
 
 export const userReducer = userSlice.reducer;
